backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
django-cors-headers==4.9.0
django-environ==0.12.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.9.0
